microservices/auth-service/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
cryptography==46.0.3
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
html5tagger==1.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.2.1
redis==7.0.1
referencing==0.37.0
rpds-py==0.28.0
sanic==25.3.0
sanic-ext==24.12.0
Sanic-Cors==2.2.0
sanic-routing==23.12.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.44
termcolor==3.2.0
tracerite==1.1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uvloop==0.22.1
websockets==15.0.1
yarl==1.22.0
factory-boy==3.3.3
aioredis==2.0.1
